fix(palette-search): add host-qualified repo and worktree resolution
Handle repo and worktree collisions in multi-host environments by
querying host-qualified maps and comparing execution host IDs when
determining current worktree status. Ensures the palette correctly
displays repo identity and marks active status for worktrees across
multiple hosts.
